@@ -161,31 +151,29 @@
         "prettier": "^2.5.1",
         "process": "^0.11.10",
         "raw-loader": "^4.0.2",
-        "react": "^17.0.2",
-        "react-ace": "^9.5.0",
-        "react-docgen": "^5.3.1",
-        "react-dom": "^17.0.2",
+        "react": "18.2.0",
+        "react-docgen": "^5.4.3",
+        "react-dom": "18.2.0",
         "react-scripts": "^5.0.1",
-        "react-test-renderer": "^17.0.2",
+        "react-test-renderer": "^18.2.0",
         "redux-mock-store": "^1.5.4",
         "sass": "^1.36.0",
         "sass-loader": "^12.1.0",
         "source-map-loader": "^3.0.0",
-        "storybook-addon-performance": "^0.16.1",
+        "style-loader": "^2.0.0",
         "terser-webpack-plugin": "^5.1.1",
         "trim": ">=0.0.3",
         "ts-loader": "^9.2.6",
-        "typescript": "4.3.2",
+        "typescript": "^4.9.5",
         "url-loader": "^4.1.1",
         "vue-style-loader": "^4.1.3",
-        "webpack": "^5.76.0",
-        "webpack-cli": "^4.5.0",
+        "webpack-cli": "^4.10.0",
         "webpack-conditional-loader": "^1.0.12",
         "webpack-dev-server": "^4.11.0"
     },
     "peerDependencies": {
-        "react": "^17",
-        "react-dom": "^17"
+        "react": "^17 || ^18",
+        "react-dom": "^17 || ^18"
     },
     "engines": {
         "node": ">=14.17",
