@@ -7,7 +7,7 @@ import {
     TableRow,
     TableBody,
     TableCell,
-} from "@material-ui/core";
+} from "@mui/material";
 import { TreeDataNode } from "@webviz/core-components";
 import cloneDeep from "lodash/cloneDeep";
 
@@ -20,7 +20,7 @@ import {
     isVariableVectorMapValid,
 } from "../../../utils/VectorCalculatorHelperFunctions";
 
-import "!style-loader!css-loader!../../../VectorCalculator.css";
+import "../../../VectorCalculator.css";
 
 interface VectorSelectorTableProps {
     vectorData: TreeDataNode[];
