@@ -1,12 +1,28 @@
-import { createStyles, makeStyles, Theme } from "@material-ui/core";
+import { styled } from "@mui/material/styles";
 import React, { useCallback, useContext, useMemo } from "react";
-import DropdownTreeSelect, { TreeNodeProps } from "react-dropdown-tree-select";
-import "!style-loader!css-loader!react-dropdown-tree-select/dist/styles.css";
+import DropdownTreeSelect, {
+    TreeNode,
+    TreeNodeProps,
+} from "react-dropdown-tree-select";
+import "react-dropdown-tree-select/dist/styles.css";
 import { useDispatch } from "react-redux";
 import { updateFilteredZones } from "../../redux/actions";
 import { Zone } from "../../redux/types";
 import { findSubzones } from "../../utils/dataUtil";
 import { DataContext } from "../DataLoader";
+
+const PREFIX = "ZoneSelector";
+
+const classes = {
+    root: `${PREFIX}-root`,
+};
+
+const StyledDropdownTreeSelect = styled(DropdownTreeSelect)(({ theme }) => ({
+    [`& .${classes.root}`]: {
+        padding: theme.spacing(1),
+        maxWidth: "250px",
+    },
+}));
 
 //Construct a stratigraphy tree as the input of react-dropdown-tree
 const extractStratigraphyTree = (stratigraphy: Zone[]): TreeNodeProps => {
@@ -58,19 +74,10 @@ export const findSelectedZones = (
     return result.map((zone) => zone.name);
 };
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        root: {
-            padding: theme.spacing(1),
-            maxWidth: "250px",
-        },
-    })
-);
 /**
  * A react component for selecting zones to display in the completions plot
  */
 const ZoneSelector: React.FC = React.memo(() => {
-    const classes = useStyles();
     // Use input data directly
     const data = useContext(DataContext);
     // Redux
@@ -82,7 +89,7 @@ const ZoneSelector: React.FC = React.memo(() => {
     );
     // Handlers
     const handleSelectionChange = useCallback(
-        (_, selectedNodes) =>
+        (_: TreeNode, selectedNodes: TreeNodeProps[]) =>
             dispatch(
                 updateFilteredZones(
                     findSelectedZones(data.stratigraphy, selectedNodes)
@@ -92,7 +99,7 @@ const ZoneSelector: React.FC = React.memo(() => {
     );
     // Render
     return (
-        <DropdownTreeSelect
+        <StyledDropdownTreeSelect
             texts={{ placeholder: "Select Zone(s)..." }}
             inlineSearchInput={true}
             showPartiallySelected={true}
