@@ -6,7 +6,11 @@ import { ExtendedLayerProps } from "../utils/layerTools";
 import { getSize } from "../wells/wellsLayer";
 import { Color } from "@deck.gl/core/typed";
 import { Feature } from "geojson";
-import { PickInfo } from "lib";
+
+// This should be fixed at some point
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type PickInfo = any;
+
 export interface BoxSelectionLayerProps<D> extends ExtendedLayerProps<D> {
     mode: string; // One of modes in MODE_MAP
     selectedFeatureIndexes: number[];
