@@ -4,7 +4,7 @@ import { MaxLengthTextField } from "./MaxLengthTextField";
 
 import { StoreActions, useStore } from "../../ExpressionsStore";
 
-import "!style-loader!css-loader!../../../VectorCalculator.css";
+import "../../../VectorCalculator.css";
 
 interface ExpressionDescriptionTextFieldProps {
     maxLength: number;
