@@ -16,7 +16,7 @@ import {
     expressionNameValidationMessage,
 } from "../../../utils/VectorCalculatorHelperFunctions";
 
-import "!style-loader!css-loader!../../../VectorCalculator.css";
+import "../../../VectorCalculator.css";
 
 type ExpressionNameTextFieldVariantType =
     | "success"
